fix(audit): resolve all -Werror warnings for GCC-13
Fixes:
- unused-but-set-variable: p_val, bytes_out, G, r, err, coeff bytes
- Wconversion: static_cast<int> for rng()%%N and static_cast<double> for .count()
- Wunused-function: [[maybe_unused]] on print_result, diagnose_scalar
- Wunused-variable: [[maybe_unused]] on g_crash
- macro redefinition: #ifndef guard for UNIFIED_AUDIT_RUNNER

Files: audit_fuzz, audit_ct, audit_integration, audit_perf,
       audit_security, unified_audit_runner, test_fuzz_parsers,
       test_musig2_frost_advanced, diag_scalar_mul

+++ audit/audit_integration.cpp
@@ -445,7 +445,7 @@ static void test_stress_mixed() {
         uint64_t v = rng();
         std::memcpy(msg.data(), &v, 8);
 
-        int op = rng() % 5;
+        int op = static_cast<int>(rng() % 5);
         switch (op) {
         case 0: { // ECDSA round-trip
             auto sig = secp256k1::ecdsa_sign(msg, sk);
